LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/response-url-expected.txt


PASS Testing response url getter with http://localhost:8800/ada 
FAIL Testing response url getter with http://localhost:8800/# assert_equals: expected "http://localhost:8800/" but got "http://localhost:8800/#"
FAIL Testing response url getter with http://localhost:8800/#ada assert_equals: expected "http://localhost:8800/" but got "http://localhost:8800/#ada"
FAIL Testing response url getter with http://localhost:8800#ada assert_equals: expected "http://localhost:8800/" but got "http://localhost:8800/#ada"
